refactor(checker): split object literal relation diagnostics packaging
