+++ crates/tsz-checker/src/lib.rs
@@ -649,6 +649,9 @@ mod nonunique_symbol_property_access_tests;
 #[path = "tests/object_literal_computed_symbol_member_tests.rs"]
 mod object_literal_computed_symbol_member_tests;
 #[cfg(test)]
+#[path = "tests/object_literal_relation_architecture_tests.rs"]
+mod object_literal_relation_architecture_tests;
+#[cfg(test)]
 #[path = "tests/object_shorthand_literal_preservation_tests.rs"]
 mod object_shorthand_literal_preservation_tests;
 #[cfg(test)]
